+++ packages/brick_rest/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Unreleased
 
-* Add `GzipHttpClient`. When instantiating the `RestProvider`, optionally compose `client` with a `GzipHttpClient` to [gzip all requests](README.md#gzipping-requests).
+* Add `GZipHttpClient`. When instantiating the `RestProvider`, optionally compose `client` with a `GZipHttpClient` to [gzip all requests](README.md#gzipping-requests).
 
 ## 0.0.7+1